Handle private/public menus with the ViewerTracker
