+++ bukkit/src/main/kotlin/com/heroslender/hmf/bukkit/screen/MenuScreen.kt
@@ -2,6 +2,7 @@ package com.heroslender.hmf.bukkit.screen
 
 import com.heroslender.hmf.bukkit.map.MapCanvas
 import com.heroslender.hmf.bukkit.screen.chunk.ScreenChunk
+import com.heroslender.hmf.bukkit.screen.tracker.MenuScreenViewerTracker
 import org.bukkit.entity.Player
 
 /**
@@ -12,6 +13,7 @@ interface MenuScreen {
     val width: Int
     val height: Int
     val chunks: Array<ScreenChunk>
+    val viewerTracker: MenuScreenViewerTracker
 
     fun holdsEntityId(entityId: Int): Boolean = chunks.any { it.id == entityId }
