+++ bukkit/src/main/kotlin/com/heroslender/hmf/bukkit/screen/tracker/PrivateMenuScreenViewerTracker.kt
@@ -1,0 +1,11 @@
+package com.heroslender.hmf.bukkit.screen.tracker
+
+import org.bukkit.entity.Player
+
+class PrivateMenuScreenViewerTracker(
+    owner: Player,
+) : MenuScreenViewerTracker {
+    override val viewers: Array<Player> = arrayOf(owner)
+
+    override fun tick() {}
+}
